requirements.txt
web3>=7.0
openai>=1.0
httpx>=0.27
python-dotenv>=1.0
pytest>=8.0
pytest-asyncio>=1.0
